Style: 팀대관 리스트 작업 (#16)
