+++ package.json
@@ -10,6 +10,7 @@
     "@types/node": "^16.18.14",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
+    "antd": "^5.3.3",
     "date-fns": "^2.29.3",
     "firebase": "^9.17.2",
     "jotai": "^2.0.3",
