+++ src/components/common/Drawer.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
 import logoImage from 'assets/logo.png';
-import { Link } from 'react-router-dom';
+import { Link, useLocation } from 'react-router-dom';
 import { AiOutlineClose } from 'react-icons/ai';
 import { useAtomValue } from 'jotai';
 import { userAtom } from 'atoms/common';
@@ -19,6 +19,12 @@ const Drawer = ({ open, closeDrawer }: Props) => {
     api.signout();
   };
 
+  const location = useLocation();
+
+  useEffect(() => {
+    closeDrawer();
+  }, [location, closeDrawer]);
+
   return (
     <>
       <Container open={open}>
@@ -30,7 +36,7 @@ const Drawer = ({ open, closeDrawer }: Props) => {
             <Logo src={logoImage} />
           </LogoLink>
           <LinkItem to="/">픽업 게임</LinkItem>
-          <LinkItem to="/">팀 대관</LinkItem>
+          <LinkItem to="/team-rental">팀 대관</LinkItem>
           <LinkItem to="/">개인 대관</LinkItem>
           <LinkItem to="/">팀 순위</LinkItem>
           <LinkItem to="/">커뮤니티</LinkItem>
